requirements.txt
asgiref==3.8.1
Django==5.1.6
django-ckeditor==6.7.0
django-cors-headers==4.7.0
django-environ==0.12.0
django-js-asset==3.0.1
djangorestframework==3.15.2
gunicorn==23.0.0
packaging==24.2
pillow==11.1.0
psycopg2==2.9.10
sqlparse==0.5.3
tzdata==2025.1
whitenoise==6.9.0
